Moved shared util code into parent util directory
