+++ tests/unit/models/test_interfaces.py
@@ -24,11 +24,10 @@ import tensorflow as tf
 import tensorflow_probability as tfp
 
 from tests.util.misc import assert_datasets_allclose, quadratic, random_seed
+from tests.util.models.models import fnc_3x_plus_10, fnc_2sin_x_over_3
 from tests.util.models.gpflow.models import (
     GaussianProcess,
     PseudoTrainableProbModel,
-    fnc_2sin_x_over_3,
-    fnc_3x_plus_10,
     gpr_model,
     sgpr_model,
 )
